requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
Pillow==10.0.1
python-dotenv==1.0.0
sqlite3
hashlib
uuid
datetime
